Articles Broken but can Push strings to Database Update 16 April
I have the functionality to push strings to database functioning. new articles are showing up in CMS, I now need to be able to store the form input as variables and use that to replace the strings... HOWEVER!!, something in this process has broken the ability to pull ALL articles from the database. not sure why.